requirements.txt
colorama==0.4.6
pillow==10.3.0
py-mon==2.0.5
pygame==2.5.2
python-dotenv==1.0.1
ttkbootstrap==1.10.1
watchdog==4.0.1
